R-stringi: Autospec creation for update from version 1.1.7 to version 1.2.2
## 1.2.2 (2018-05-01) **CRAN**

* [GENERAL] #193: `stringi` is now bundled with ICU4C 61.1,
which is used on most Windows and OS X builds as well as on *nix systems
not equipped with ICU. However, if the C++11 support is disabled,
stringi will be built against ICU4C 55.1. The update to ICU brings
Unicode 10.0 support, including new emoji characters.

* [BUGFIX] #288: stri_match did not return the correct number of columns
when input was empty.

* [NEW FEATURE] #188: `stri_enc_detect` now returns a list of data frames.

* [NEW FEATURE] #289: `stri_flatten` gained `na_empty` `omit_empty` arguments.

(NEWS truncated at 15 lines)